Revert "feat: 🎸 rollovers on create balance"
