+++ vite/src/views/customers2/components/sheets/BalanceCreateSheet.tsx
@@ -1,10 +1,4 @@
-import {
-	type Feature,
-	FeatureType,
-	isContUseFeature,
-	ResetInterval,
-	type RolloverConfig,
-} from "@autumn/shared";
+import { type Feature, FeatureType, ResetInterval } from "@autumn/shared";
 import { CalendarXIcon, InfinityIcon } from "@phosphor-icons/react";
 import { useState } from "react";
 import { toast } from "sonner";
@@ -32,7 +26,6 @@ import { useSheetStore } from "@/hooks/stores/useSheetStore";
 import { useAxiosInstance } from "@/services/useAxiosInstance";
 import { getBackendErr } from "@/utils/genUtils";
 import { useCusQuery } from "@/views/customers/customer/hooks/useCusQuery";
-import { RolloverConfigForm } from "@/views/products/plan/components/edit-plan-feature/advanced-settings/RolloverConfigForm";
 import { useCustomerContext } from "../../customer/CustomerContext";
 
 const RESET_INTERVAL_LABELS: Record<string, string> = {
@@ -61,11 +54,6 @@ export function BalanceCreateSheet() {
 	const [resetInterval, setResetInterval] = useState<string>("");
 	const [oneOff, setOneOff] = useState(false);
 	const [expiresAt, setExpiresAt] = useState<number | null>(null);
-	const [rollover, setRollover] = useState<RolloverConfig | null>(null);
-
-	// Rollover requires a recurring reset interval. If the interval goes away
-	// (or the user picks one-off/unlimited), clear rollover to keep state valid.
-	const hasRecurringReset = Boolean(resetInterval) && !oneOff && !unlimited;
 
 	const nonArchivedFeatures = features.filter((f: Feature) => !f.archived);
 	const selectedFeature = nonArchivedFeatures.find(
@@ -74,36 +62,6 @@ export function BalanceCreateSheet() {
 	const isMetered =
 		selectedFeature?.type === FeatureType.Metered ||
 		selectedFeature?.type === FeatureType.CreditSystem;
-
-	// Continuous-use features (non-consumable) don't support rollover — the
-	// server rejects it in validateCreateBalanceParams, so we gate the UI too.
-	const isContUse = selectedFeature
-		? isContUseFeature({ feature: selectedFeature })
-		: false;
-	const canEnableRollover = isMetered && !isContUse && hasRecurringReset;
-
-	/**
-	 * Authoritative feature-select handler.
-	 *
-	 * Rollover is only valid for metered/credit-system features that are
-	 * consumable (non-continuous). When the user swaps features, clear any
-	 * existing rollover state synchronously here — no useEffect needed.
-	 */
-	const handleSelectFeature = (nextFeatureId: string) => {
-		setFeatureId(nextFeatureId);
-		const nextFeature = nonArchivedFeatures.find(
-			(f: Feature) => f.id === nextFeatureId,
-		);
-		const nextIsMetered =
-			nextFeature?.type === FeatureType.Metered ||
-			nextFeature?.type === FeatureType.CreditSystem;
-		const nextIsContUse = nextFeature
-			? isContUseFeature({ feature: nextFeature })
-			: false;
-		if (!nextIsMetered || nextIsContUse) {
-			setRollover(null);
-		}
-	};
 
 	const handleCreate = async () => {
 		const customerId = customer?.id || customer?.internal_id;
@@ -137,10 +95,6 @@ export function BalanceCreateSheet() {
 			params.reset = { interval: resetInterval };
 		}
 
-		if (rollover && canEnableRollover) {
-			params.rollover = rollover;
-		}
-
 		if (expiresAt) {
 			params.expires_at = expiresAt;
 		}
@@ -171,7 +125,7 @@ export function BalanceCreateSheet() {
 					<FeatureSearchDropdown
 						features={nonArchivedFeatures}
 						value={featureId || null}
-						onSelect={handleSelectFeature}
+						onSelect={setFeatureId}
 					/>
 				</SheetSection>
 
@@ -210,7 +164,6 @@ export function BalanceCreateSheet() {
 													setIncludedGrant("");
 													setResetInterval("");
 													setOneOff(false);
-													setRollover(null);
 												}
 											}}
 											className="py-1 w-26 text-t4 gap-2"
@@ -251,10 +204,7 @@ export function BalanceCreateSheet() {
 											checked={oneOff}
 											onCheckedChange={(checked) => {
 												setOneOff(checked);
-												if (checked) {
-													setResetInterval("");
-													setRollover(null);
-												}
+												if (checked) setResetInterval("");
 											}}
 											className="py-1 w-26 text-t4 gap-2 justify-start"
 										>
@@ -274,14 +224,6 @@ export function BalanceCreateSheet() {
 										use24Hour
 									/>
 								</div>
-							)}
-
-							{isMetered && !unlimited && !isContUse && (
-								<RolloverConfigForm
-									value={rollover}
-									onChange={setRollover}
-									disabled={!canEnableRollover}
-								/>
 							)}
 						</div>
 					</SheetSection>
